requirements.txt
streamlit>=1.22.0
tensorflow>=2.12.0
numpy>=1.21.0
matplotlib>=3.5.0
Pillow>=9.0.0
opencv-python>=4.5.0
pandas>=1.3.0
scikit-learn>=1.0.0
seaborn>=0.11.0
plotly>=5.13.0
scipy>=1.7.0 